+++ packages/api/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @zitadel/api
 
+## 0.1.0-alpha.17
+
 ## 0.1.0-alpha.16
 
 ### Minor Changes